+++ CANCHASBA-API/src/main/java/com/app/CANCHASBA_API/models/dao/CanchaDao.java
@@ -3,5 +3,5 @@ package com.app.CANCHASBA_API.models.dao;
 import com.app.CANCHASBA_API.models.entity.Cancha;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CanchaDao extends CrudRepository<Cancha,Integer> {
+public interface CanchaDao extends CrudRepository<Cancha,Long> {
 }
